move `blobToBase64` definition
