+++ src/DashboardUI/src/pages/application/estimation-tool/EstimationToolPage.tsx
@@ -17,7 +17,7 @@ import { toast } from 'react-toastify';
 import { ApplicationNavbar } from '../components/ApplicationNavbar';
 import { useMapContext } from '../../../contexts/MapProvider';
 import { MapPolygon } from '../../../data-contracts/MapPolygon';
-import { blobToBase64 } from '../../../utilities/blobUtility';
+import { blobToBase64 } from '../../../utilities/fileUploadHelpers';
 
 import './estimation-tool-page.scss';
