+++ src/DashboardUI/src/utilities/fileUploadHelpers.ts
@@ -34,3 +34,13 @@ export const downloadFilesFromBlobStorage = async (sasToken: string, fileName: s
   document.body.appendChild(link);
   link.click();
 };
+
+export const blobToBase64 = (blob: Blob): Promise<string> => {
+  const reader = new FileReader();
+  reader.readAsDataURL(blob);
+  return new Promise((resolve) => {
+    reader.onloadend = () => {
+      resolve(reader.result as string);
+    };
+  });
+};
